Customer model updated
https://github.com/opencart/opencart/commit/fa7b453e6e33f5971c593970005ef68719206197

+++ upload/admin/model/customer/customer.php
@@ -437,10 +437,10 @@ class ModelCustomerCustomer extends Model {
 		return (int)$query->row['total'];
 	}
 
-	public function getTransactionTotal(int $customer_id): int {
+	public function getTransactionTotal(int $customer_id): float {
 		$query = $this->db->query("SELECT SUM(amount) AS total FROM " . DB_PREFIX . "customer_transaction WHERE customer_id = '" . (int)$customer_id . "'");
 
-		return (int)$query->row['total'];
+		return (float)$query->row['total'];
 	}
 
 	public function getTotalTransactionsByOrderId(int $order_id): int {
